django/requirements.txt
asgiref==3.6.0
Django==4.1.5
django-cors-headers==3.12.0
django-filter==22.1
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2==2.9.3
pytz==2022.7.1
sqlparse==0.4.3
